+++ src/app/services/authentication/authentication.service.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { NotificationsService } from '../notifications/notifications.service';
 import { LoadingService } from '../loading/loading.service';
-import { UserService } from '../user.service';
+import { UserService } from '../user/user.service';
 import { Kinvey } from 'kinvey-angular2-sdk';
 
 import {
