+++ src/app/services/index.ts
@@ -1,8 +1,8 @@
 export * from './api.constants';
-export * from './movies.service';
-export * from './search.service';
+export * from './movies/movies.service';
+export * from './search/search.service';
 export * from './authentication/authentication.service';
-export * from './user.service';
+export * from './user/user.service';
 export * from './resolvers/movie-details.resolver';
 export * from './guards/user-guard.service';
 export * from './notifications/notifications.service';
