+++ src/app/services/resolvers/movie-details.resolver.ts
@@ -5,7 +5,7 @@ import {
   ActivatedRouteSnapshot,
   RouterStateSnapshot
 } from '@angular/router';
-import { MoviesService } from '../movies.service';
+import { MoviesService } from '../movies/movies.service';
 
 @Injectable()
 export class MovieDetailsResolver implements Resolve<object> {
